app/backend/app/requirements.txt
Flask
mysql-connector
Flask-Cors